@@ -3,10 +3,7 @@ package com.peopleofandroido.chillaxingcat
 import android.app.Application
 import com.peopleofandroido.base.di.componentModule
 import com.peopleofandroido.base.di.networkModule
-import com.peopleofandroido.chillaxingcat.di.appViewModelModule
-import com.peopleofandroido.chillaxingcat.di.databaseModule
-import com.peopleofandroido.chillaxingcat.di.repositoryModule
-import com.peopleofandroido.chillaxingcat.di.useCaseModule
+import com.peopleofandroido.chillaxingcat.di.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -18,15 +15,16 @@ class ChillaxingCatApplication : Application() {
         startKoin {
             androidLogger() //androidLogger를 KoinLogger로 사용
             androidContext(this@ChillaxingCatApplication) // 해당 안드로이드 context 사용
-                //koin.loadModules(arrayListOf(...)) //이렇게로도 가능
-            modules(
+            koin.loadModules(arrayListOf( //이렇게로도 가능
+//            modules(
                 appViewModelModule,
                 componentModule,
                 repositoryModule,
                 databaseModule,
                 useCaseModule,
                 networkModule,
-            )
+                appDataStoreModule,
+            ))
         }
     }
 }
