@@ -6,7 +6,7 @@ import org.koin.dsl.module
 
 val useCaseModule = module {
     single {
-        UseCases(get(), get(), get(), get(), get(), get(), get(), get(), get())
+        UseCases(get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get())
     }
 
     single { AddDayOff(get(), get()) }
@@ -19,4 +19,7 @@ val useCaseModule = module {
     single { AddRestingTime(get(), get()) }
     single { EditRestingTime(get(), get()) }
     single { GetRestingTime(get(), get()) }
+    single { FindOutRestingDaysInMonth(get(), get()) }
+    single { GetNotificationStatus(get(), get()) }
+    single { PutNotificationStatus(get(), get()) }
 }
