@@ -2,12 +2,11 @@ package com.peopleofandroido.chillaxingcat.presentation
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
-import androidx.navigation.fragment.findNavController
+import androidx.navigation.findNavController
 import com.peopleofandroido.base.common.BaseBindingActivity
 import com.peopleofandroido.base.common.NavManager
 import com.peopleofandroido.chillaxingcat.R
 import com.peopleofandroido.chillaxingcat.databinding.ActivityMainBinding
-import kotlinx.android.synthetic.main.activity_main.*
 import org.koin.android.ext.android.get
 
 class MainActivity : BaseBindingActivity<ActivityMainBinding>() {
@@ -16,7 +15,7 @@ class MainActivity : BaseBindingActivity<ActivityMainBinding>() {
         R.layout.activity_main
 
     private val navController
-        get() = nav_host.findNavController()
+        get() = binding.navHost.findNavController()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
